Add a test for Bug #14561
Since it takes 5 seconds to run, I split this into a separate commit
that maybe we don't merge into trunk.

I also question the value of this as a regression test since the
problem needs a specific build configuration and also relies on
libpthread's implementation (particularly what it does with the
stack of a native thread that has exited).