Clarify Kernel#raise signature and add a spec for it
The documentation states that the `string` argument must be a `String`
instance, but that's not quite true.

`Kernel#raise` itself doesn't enforce the `message` type,
it's the exception's initializer that does.

For instance the following code works just fine, and can be seen
in popular gems.

```
class ClientError < StandardError
  attr_reader :http_response

  def initialize(http_response)
    @http_response = http_response
    super("some dynamic message")
  end
end

raise SomeErrorClass, http_response
```