make sure we update the allocated objects stats
otherwise the count we get from walking the heap disagrees with the
stats.

I'm in two minds about this patch, it does fix the tests, however it
depends on whether `total_allocated_objects` refers to ruby objects or
heap slots. whilst the heap slots are technically allocated (they're
filled with garbage) eventually the contents of this will be the
associated with an RVALUE and could be considered to be part of that
object.

so I wasn't sure whether to either to do what I've done in this patch
(make the stats also take into account the garbage) or whether to make
the heap walking algorigthm ignore garbage...

I suspect that in the long term the latter approach will work better,
but this is a much smaller change, and fixes my test.