Remove write barrier exemption for T_ICLASS
This is a backport of 264e4cd04fbcdcb739a1ff9a84e19afe66005cb2.
Before this commit, iclasses were "shady", or not protected by write
barriers. Because of that, the GC needs to spend more time marking these
objects than otherwise.

Applications that make heavy use of modules should see reduction in GC
time as they have a significant number of live iclasses on the heap.