Perform full sweep when preparing a new heap free page
This code is called when creating a new object on the slow path and
there isn't a free region large enough to hold the object and it's
payload.

It first attempts to run the GC to free up pages on an existing page
before allocating a new one. When it starts the GC it does it in
incremental sweeping mode. [This
commit](https://github.com/Shopify/ruby/commit/0bebea985dd1b759973e285e87ed9dfa8610d253#diff-d1cee85c3b0e24a64519c11150abe26fd0b5d8628a23d356dd0b535ac4595d49)
adds an assertion and a comment informing us that adding a page to the
eden heap during incremental sweeping is forbidden.

This commit provides a quick (but non-ideal) patch that makes sure that
when preparing a new free page we always run a full sweep. This has
performance implications to object creation when the free list is
running out of space or fragmented.

But importantly, this allows me to compile the code after rebasing this
branch on top of master