Associate `initialize` CI with opt_new in HIR
This commit saves the CI for `initialize` calls that correspond to
`opt_new` instructions.  ZJIT treats `opt_new` as "always taking the slow path"
and later inlines the call to `initialize`.  When we add the call to
`initialize` we need the call info object for that method call.  This
commit saves the CI for later retrieval